Predict vulpkanin scent tracking verb and actions. (#4394)
* scent tracker is now predicted, yippee

* Optimize performance (moved trycomp)
